优化小图片下载流程
小图片的下载不再通过CoreEngine中转，直接调用UIImageView的分类Cache中的方法即可

+++ ExceedMVC/Core/DB/DBBase/DBStatement.m
@@ -84,7 +84,7 @@
     sqlite3_bind_int64(_stmt, index, value);
 }
 - (void)bindData:(NSData*)value forIndex:(int)index {
-    sqlite3_bind_blob(_stmt, index, value.bytes, value.length, SQLITE_TRANSIENT);
+    sqlite3_bind_blob(_stmt, index, value.bytes, (int)value.length, SQLITE_TRANSIENT);
 }
 - (void)bindBool:(BOOL)value forIndex:(int)index {
 	sqlite3_bind_int(_stmt, index, value?1:0);
